refactor(subscription): split subscriptionService into focused modules
Extract self-contained pieces of the 1791-line subscriptionService.ts into
services/subscription/: types (Subscription), constants (Twitch limits),
helpers (notify/error/log/filename-template), youtubeAuthor (channel
name resolution), and twitchSubscription (the full Twitch polling/
processing cluster as free functions).

SubscriptionService stays the public facade: it delegates to the extracted
functions and keeps thin instance methods for checkTwitchSubscription and
processTwitchSubscriptionVideos, which existing tests spy on / call
directly. Behavior is unchanged; the entry file drops to ~1160 lines.
