+++ backend/src/services/subscription/constants.ts
@@ -1,0 +1,18 @@
+export const MAX_TWITCH_SUBSCRIPTION_PAGES_PER_CHECK = 5;
+export const MAX_TWITCH_SUBSCRIPTION_SCANNED_VIDEOS = 500;
+export const MAX_TWITCH_SUBSCRIPTION_DOWNLOADS_PER_CHECK = Math.max(
+  Number.parseInt(
+    process.env.TWITCH_SUBSCRIPTION_MAX_DOWNLOADS_PER_CHECK || "3",
+    10
+  ) || 3,
+  1
+);
+export const RETRYABLE_TWITCH_API_ERROR_CODES = new Set([
+  "ECONNABORTED",
+  "ECONNREFUSED",
+  "ECONNRESET",
+  "ENOTFOUND",
+  "ETIMEDOUT",
+  "EAI_AGAIN",
+  "ERR_NETWORK",
+]);
