+++ backend/src/services/subscription/types.ts
@@ -1,0 +1,36 @@
+export interface Subscription {
+  id: string;
+  author: string;
+  authorUrl: string;
+  interval: number;
+  lastVideoLink?: string;
+  lastCheck?: number;
+  downloadCount: number;
+  createdAt: number;
+
+  platform: string;
+  paused?: number;
+
+  // Playlist subscription fields
+  playlistId?: string;
+  playlistTitle?: string;
+  subscriptionType?: string; // 'author' or 'playlist'
+  collectionId?: string;
+
+  // Shorts support
+  downloadShorts?: number; // 0 or 1
+  lastShortVideoLink?: string;
+
+  // Twitch support
+  twitchBroadcasterId?: string;
+  twitchBroadcasterLogin?: string;
+  lastTwitchVideoId?: string;
+
+  // Retention
+  retentionDays?: number | null;
+
+  // Statistics-related durable state
+  consecutiveFailureCount?: number;
+  lastCheckStatus?: string | null;
+  lastFailureReason?: string | null;
+}
