+++ backend/src/services/subscriptionService.ts
@@ -23,173 +23,27 @@ import {
     downloadYouTubeVideo,
 } from "./downloadService";
 import { BilibiliDownloader } from "./downloaders/BilibiliDownloader";
-import {
-  getTwitchChannelVideos,
-  TwitchYtDlpVideoEntry,
-} from "./downloaders/ytdlp/ytdlpTwitch";
+import { getTwitchChannelVideos } from "./downloaders/ytdlp/ytdlpTwitch";
 import { YtDlpDownloader } from "./downloaders/YtDlpDownloader";
-import { FilenameTemplateSourceOptions } from "./filenameTemplate/types";
-import { recordEvent, bucketDownloadError, platformFromUrl } from "./statistics";
+import { recordEvent, bucketDownloadError } from "./statistics";
 import * as storageService from "./storageService";
 import { runSubscriptionRetentionCleanup } from "./subscriptionRetentionService";
-import { TelegramService } from "./telegramService";
-import { TwitchVideoInfo, twitchApiService } from "./twitchService";
+import { twitchApiService } from "./twitchService";
+import {
+  buildFilenameTemplateSourceOptions,
+  getErrorMessage,
+  getSubscriptionLogContext,
+  notifySubscriptionDownloadResult,
+} from "./subscription/helpers";
+import { Subscription } from "./subscription/types";
+import { resolveYouTubeAuthorName } from "./subscription/youtubeAuthor";
+import {
+  checkTwitchSubscription as checkTwitchSubscriptionImpl,
+  isEligibleTwitchVideo,
+  processTwitchSubscriptionVideos as processTwitchSubscriptionVideosImpl,
+} from "./subscription/twitchSubscription";
 
-const MAX_TWITCH_SUBSCRIPTION_PAGES_PER_CHECK = 5;
-const MAX_TWITCH_SUBSCRIPTION_SCANNED_VIDEOS = 500;
-const MAX_TWITCH_SUBSCRIPTION_DOWNLOADS_PER_CHECK = Math.max(
-  Number.parseInt(
-    process.env.TWITCH_SUBSCRIPTION_MAX_DOWNLOADS_PER_CHECK || "3",
-    10
-  ) || 3,
-  1
-);
-const RETRYABLE_TWITCH_API_ERROR_CODES = new Set([
-  "ECONNABORTED",
-  "ECONNREFUSED",
-  "ECONNRESET",
-  "ENOTFOUND",
-  "ETIMEDOUT",
-  "EAI_AGAIN",
-  "ERR_NETWORK",
-]);
-
-function shouldFallbackToTwitchYtDlp(error: unknown): boolean {
-  if (error instanceof ValidationError) {
-    return (
-      error.field === "twitchClientId" || error.field === "twitchClientSecret"
-    );
-  }
-
-  if (error && typeof error === "object") {
-    const errorWithResponse = error as {
-      code?: unknown;
-      message?: unknown;
-      request?: unknown;
-      response?: { status?: unknown };
-    };
-
-    if (typeof errorWithResponse.response?.status === "number") {
-      return true;
-    }
-
-    if (
-      typeof errorWithResponse.code === "string" &&
-      RETRYABLE_TWITCH_API_ERROR_CODES.has(errorWithResponse.code)
-    ) {
-      return true;
-    }
-
-    if (errorWithResponse.request !== undefined) {
-      return true;
-    }
-  }
-
-  return (
-    error instanceof Error &&
-    error.message.includes("Twitch API is temporarily rate limited")
-  );
-}
-
-function notifySubscriptionDownloadResult(context: {
-  taskTitle: string;
-  status: "success" | "fail";
-  sourceUrl?: string;
-  error?: string;
-}): void {
-  void TelegramService.notifyTaskComplete(context).catch((error) => {
-    logger.error(
-      "Subscription Telegram notification failed:",
-      error instanceof Error ? error : new Error(String(error))
-    );
-  });
-}
-
-function getErrorMessage(error: unknown, fallback: string): string {
-  if (error instanceof Error) {
-    return error.message;
-  }
-
-  if (error && typeof error === "object") {
-    const maybeError = error as { message?: unknown };
-    if (typeof maybeError.message === "string") {
-      return maybeError.message;
-    }
-  }
-
-  return fallback;
-}
-
-function buildFilenameTemplateSourceOptions(
-  sub: Subscription,
-  mediaPlaylistIndex?: number
-): FilenameTemplateSourceOptions {
-  const isPlaylist =
-    sub.subscriptionType === "playlist" || Boolean(sub.playlistId);
-
-  return {
-    sourceCustomName: sub.author,
-    sourceCollectionName: sub.playlistTitle || sub.author,
-    sourceCollectionId: sub.playlistId || sub.collectionId || "",
-    sourceCollectionType: isPlaylist ? "playlist" : "channel",
-    mediaPlaylistIndex,
-  };
-}
-
-export interface Subscription {
-  id: string;
-  author: string;
-  authorUrl: string;
-  interval: number;
-  lastVideoLink?: string;
-  lastCheck?: number;
-  downloadCount: number;
-  createdAt: number;
-
-  platform: string;
-  paused?: number;
-
-  // Playlist subscription fields
-  playlistId?: string;
-  playlistTitle?: string;
-  subscriptionType?: string; // 'author' or 'playlist'
-  collectionId?: string;
-
-  // Shorts support
-  downloadShorts?: number; // 0 or 1
-  lastShortVideoLink?: string;
-
-  // Twitch support
-  twitchBroadcasterId?: string;
-  twitchBroadcasterLogin?: string;
-  lastTwitchVideoId?: string;
-
-  // Retention
-  retentionDays?: number | null;
-
-  // Statistics-related durable state
-  consecutiveFailureCount?: number;
-  lastCheckStatus?: string | null;
-  lastFailureReason?: string | null;
-}
-
-function getSubscriptionLogContext(
-  sub: {
-    id: string;
-    author?: string | null;
-    authorUrl?: string | null;
-    platform?: string | null;
-  },
-  extras: Record<string, unknown> = {}
-): Record<string, unknown> {
-  return {
-    subscriptionId: sub.id,
-    author: sub.author,
-    authorUrl: sub.authorUrl,
-    platform: sub.platform,
-    ...extras,
-  };
-}
+export type { Subscription } from "./subscription/types";
 
 export class SubscriptionService {
   private static instance: SubscriptionService;
@@ -247,114 +101,7 @@ export class SubscriptionService {
 
       // If author name not provided, try to get it from channel URL using yt-dlp
       if (!providedAuthorName) {
-        try {
-          const {
-            executeYtDlpJson,
-            getNetworkConfigFromUserConfig,
-            getUserYtDlpConfig,
-          } = await import("../utils/ytDlpUtils");
-          const userConfig = getUserYtDlpConfig(authorUrl);
-          const networkConfig = getNetworkConfigFromUserConfig(userConfig);
-
-          // Construct URL to get videos from the channel
-          let targetUrl = authorUrl;
-          if (
-            !targetUrl.includes("/videos") &&
-            !targetUrl.includes("/shorts") &&
-            !targetUrl.includes("/streams")
-          ) {
-            // Append /videos to get the videos playlist
-            if (targetUrl.endsWith("/")) {
-              targetUrl = `${targetUrl}videos`;
-            } else {
-              targetUrl = `${targetUrl}/videos`;
-            }
-          }
-
-          // Try to get channel info from the channel URL
-          const info = await executeYtDlpJson(targetUrl, {
-            ...networkConfig,
-            noWarnings: true,
-            flatPlaylist: true,
-            playlistEnd: 1,
-          });
-
-          // Try to get uploader/channel name from the first video or channel info
-          if (info.uploader) {
-            authorName = info.uploader;
-          } else if (info.channel) {
-            authorName = info.channel;
-          } else if (
-            info.channel_id &&
-            info.entries &&
-            info.entries.length > 0
-          ) {
-            // If we have entries, try to get info from the first video
-            const firstVideo = info.entries[0];
-            if (firstVideo && firstVideo.url) {
-              try {
-                const videoInfo = await executeYtDlpJson(firstVideo.url, {
-                  ...networkConfig,
-                  noWarnings: true,
-                });
-                if (videoInfo.uploader) {
-                  authorName = videoInfo.uploader;
-                } else if (videoInfo.channel) {
-                  authorName = videoInfo.channel;
-                }
-              } catch (videoError) {
-                logger.error(
-                  "Error fetching video info for channel name:",
-                  videoError
-                );
-              }
-            }
-          }
-
-          // Fallback: try to extract from URL if still not found
-          if (
-            authorName === "Unknown Author" ||
-            authorName === providedAuthorName
-          ) {
-            const match = decodeURI(authorUrl).match(/youtube\.com\/(@[^\/]+)/);
-            if (match && match[1]) {
-              authorName = match[1];
-            } else {
-              const parts = authorUrl.split("/");
-              if (parts.length > 0) {
-                const lastPart = parts[parts.length - 1];
-                if (
-                  lastPart &&
-                  lastPart !== "videos" &&
-                  lastPart !== "about" &&
-                  lastPart !== "channel"
-                ) {
-                  authorName = lastPart;
-                }
-              }
-            }
-          }
-        } catch (error) {
-          logger.error("Error fetching YouTube channel info:", error);
-          // Fallback: try to extract from URL
-          const match = decodeURI(authorUrl).match(/youtube\.com\/(@[^\/]+)/);
-          if (match && match[1]) {
-            authorName = match[1];
-          } else {
-            const parts = authorUrl.split("/");
-            if (parts.length > 0) {
-              const lastPart = parts[parts.length - 1];
-              if (
-                lastPart &&
-                lastPart !== "videos" &&
-                lastPart !== "about" &&
-                lastPart !== "channel"
-              ) {
-                authorName = lastPart;
-              }
-            }
-          }
-        }
+        authorName = await resolveYouTubeAuthorName(authorUrl, providedAuthorName);
       }
     } else if (isTwitchChannelUrl(authorUrl)) {
       authorUrl = normalizeTwitchChannelUrl(authorUrl);
@@ -382,7 +129,7 @@ export class SubscriptionService {
           }
         );
         const newestEligibleVideo = videos.find((video) =>
-          this.isEligibleTwitchVideo(video)
+          isEligibleTwitchVideo(video)
         );
 
         authorName = channel.displayName || providedAuthorName || channel.login;
@@ -1306,252 +1053,15 @@ export class SubscriptionService {
     }
   }
 
-  private isEligibleTwitchVideo(video: TwitchVideoInfo): boolean {
-    return video.type === "archive" || video.type === "upload";
+  /**
+   * Thin delegations to the extracted Twitch implementation. Retained as
+   * instance methods because tests spy on / exercise them directly.
+   */
+  private checkTwitchSubscription(sub: Subscription): Promise<number> {
+    return checkTwitchSubscriptionImpl(sub);
   }
 
-  private async checkTwitchSubscription(sub: Subscription): Promise<number> {
-    const now = Date.now();
-    const lockResult = await db
-      .update(subscriptions)
-      .set({ lastCheck: now })
-      .where(eq(subscriptions.id, sub.id))
-      .returning({ id: subscriptions.id });
-
-    if (lockResult.length === 0) {
-      logger.warn(
-        "Twitch subscription was deleted before polling",
-        getSubscriptionLogContext(sub)
-      );
-      return 0;
-    }
-
-    if (!twitchApiService.isConfigured()) {
-      return await this.checkTwitchSubscriptionWithYtDlp(sub);
-    }
-
-    try {
-      return await this.checkTwitchSubscriptionWithApi(sub);
-    } catch (error) {
-      if (!shouldFallbackToTwitchYtDlp(error)) {
-        throw error;
-      }
-
-      logger.warn(
-        "Falling back to yt-dlp for Twitch subscription after Helix polling failed",
-        error instanceof Error ? error : new Error(String(error)),
-        getSubscriptionLogContext(sub)
-      );
-      return await this.checkTwitchSubscriptionWithYtDlp(sub);
-    }
-  }
-
-  private async checkTwitchSubscriptionWithApi(
-    sub: Subscription
-  ): Promise<number> {
-    twitchApiService.ensureConfigured();
-
-    let channel = sub.twitchBroadcasterId
-      ? await twitchApiService.getChannelById(sub.twitchBroadcasterId)
-      : null;
-
-    if (!channel) {
-      const channelLogin =
-        sub.twitchBroadcasterLogin || extractTwitchChannelLogin(sub.authorUrl);
-      if (!channelLogin) {
-        throw new ValidationError(
-          `Invalid Twitch channel URL: ${sub.authorUrl}`,
-          "authorUrl"
-        );
-      }
-      channel = await twitchApiService.getChannelByLogin(channelLogin);
-    }
-
-    if (!channel) {
-      logger.warn(
-        `Twitch channel for subscription ${sub.id} could not be resolved`
-      );
-      return 0;
-    }
-
-    await db
-      .update(subscriptions)
-      .set({
-        author: channel.displayName,
-        authorUrl: channel.url,
-        twitchBroadcasterId: channel.id,
-        twitchBroadcasterLogin: channel.login,
-      })
-      .where(eq(subscriptions.id, sub.id));
-
-    const unseenVideos: TwitchVideoInfo[] = [];
-    let cursor: string | undefined;
-    let pagesFetched = 0;
-    let scannedVideos = 0;
-    let foundMarker = false;
-
-    while (
-      pagesFetched < MAX_TWITCH_SUBSCRIPTION_PAGES_PER_CHECK &&
-      scannedVideos < MAX_TWITCH_SUBSCRIPTION_SCANNED_VIDEOS
-    ) {
-      const response = await twitchApiService.listVideosByBroadcaster(
-        channel.id,
-        {
-          after: cursor,
-          first: 100,
-          type: "all",
-        }
-      );
-      pagesFetched += 1;
-
-      if (response.videos.length === 0) {
-        break;
-      }
-
-      for (const video of response.videos) {
-        scannedVideos += 1;
-
-        if (!this.isEligibleTwitchVideo(video)) {
-          if (scannedVideos >= MAX_TWITCH_SUBSCRIPTION_SCANNED_VIDEOS) {
-            break;
-          }
-          continue;
-        }
-
-        if (sub.lastTwitchVideoId && video.id === sub.lastTwitchVideoId) {
-          foundMarker = true;
-          break;
-        }
-
-        unseenVideos.push(video);
-        if (scannedVideos >= MAX_TWITCH_SUBSCRIPTION_SCANNED_VIDEOS) {
-          break;
-        }
-      }
-
-      if (
-        foundMarker ||
-        !response.cursor ||
-        scannedVideos >= MAX_TWITCH_SUBSCRIPTION_SCANNED_VIDEOS
-      ) {
-        break;
-      }
-
-      cursor = response.cursor;
-    }
-
-    if (unseenVideos.length === 0) {
-      return 0;
-    }
-
-    return await this.processTwitchSubscriptionVideos(
-      sub,
-      unseenVideos
-        .reverse()
-        .slice(0, MAX_TWITCH_SUBSCRIPTION_DOWNLOADS_PER_CHECK)
-        .map((video) => ({
-        id: video.id,
-        url: video.url,
-        title: video.title,
-        authorName: video.userName || channel.displayName,
-      }))
-    );
-  }
-
-  private async checkTwitchSubscriptionWithYtDlp(
-    sub: Subscription
-  ): Promise<number> {
-    const fallbackLogin =
-      sub.twitchBroadcasterLogin || extractTwitchChannelLogin(sub.authorUrl);
-    if (!fallbackLogin) {
-      throw new ValidationError(
-        `Invalid Twitch channel URL: ${sub.authorUrl}`,
-        "authorUrl"
-      );
-    }
-
-    const normalizedUrl = normalizeTwitchChannelUrl(sub.authorUrl);
-    const unseenVideos: TwitchYtDlpVideoEntry[] = [];
-    let pagesFetched = 0;
-    let scannedVideos = 0;
-    let foundMarker = false;
-    let resolvedAuthor = sub.author;
-    let resolvedLogin = sub.twitchBroadcasterLogin || fallbackLogin;
-
-    while (
-      pagesFetched < MAX_TWITCH_SUBSCRIPTION_PAGES_PER_CHECK &&
-      scannedVideos < MAX_TWITCH_SUBSCRIPTION_SCANNED_VIDEOS
-    ) {
-      const response = await getTwitchChannelVideos(normalizedUrl, {
-        startIndex: pagesFetched * 100,
-        limit: 100,
-      });
-      pagesFetched += 1;
-
-      if (response.channelName) {
-        resolvedAuthor = response.channelName;
-      }
-      if (response.channelLogin) {
-        resolvedLogin = response.channelLogin;
-      }
-
-      if (pagesFetched === 1) {
-        await db
-          .update(subscriptions)
-          .set({
-            author: resolvedAuthor,
-            authorUrl: normalizedUrl,
-            twitchBroadcasterLogin: resolvedLogin,
-          })
-          .where(eq(subscriptions.id, sub.id));
-      }
-
-      if (response.videos.length === 0) {
-        break;
-      }
-
-      for (const video of response.videos) {
-        scannedVideos += 1;
-
-        if (sub.lastTwitchVideoId && video.id === sub.lastTwitchVideoId) {
-          foundMarker = true;
-          break;
-        }
-
-        unseenVideos.push(video);
-        if (scannedVideos >= MAX_TWITCH_SUBSCRIPTION_SCANNED_VIDEOS) {
-          break;
-        }
-      }
-
-      if (
-        foundMarker ||
-        response.videos.length < 100 ||
-        scannedVideos >= MAX_TWITCH_SUBSCRIPTION_SCANNED_VIDEOS
-      ) {
-        break;
-      }
-    }
-
-    if (unseenVideos.length === 0) {
-      return 0;
-    }
-
-    return await this.processTwitchSubscriptionVideos(
-      sub,
-      unseenVideos
-        .reverse()
-        .slice(0, MAX_TWITCH_SUBSCRIPTION_DOWNLOADS_PER_CHECK)
-        .map((video) => ({
-        id: video.id,
-        url: video.url,
-        title: video.title,
-        authorName: video.author || resolvedAuthor,
-      }))
-    );
-  }
-
-  private async processTwitchSubscriptionVideos(
+  private processTwitchSubscriptionVideos(
     sub: Subscription,
     videosToProcess: Array<{
       id: string;
@@ -1560,142 +1070,7 @@ export class SubscriptionService {
       authorName?: string | null;
     }>
   ): Promise<number> {
-    let currentLastVideoLink = sub.lastVideoLink || "";
-    let currentLastTwitchVideoId = sub.lastTwitchVideoId;
-    let currentDownloadCount = sub.downloadCount || 0;
-    let newVideoCount = 0;
-
-    for (const video of videosToProcess) {
-      const existingDownload = storageService.checkVideoDownloadBySourceId(
-        video.id,
-        "twitch"
-      );
-
-      if (existingDownload.found) {
-        currentLastTwitchVideoId = video.id;
-        currentLastVideoLink = video.url;
-
-        await db
-          .update(subscriptions)
-          .set({
-            lastTwitchVideoId: currentLastTwitchVideoId,
-            lastVideoLink: currentLastVideoLink,
-          })
-          .where(eq(subscriptions.id, sub.id));
-        continue;
-      }
-
-      let twitchVideoDownloaded = false;
-      let downloadedTwitchTitle = video.title || `Video from ${sub.author}`;
-      try {
-        const downloadResult = await downloadYouTubeVideo(
-          video.url,
-          undefined,
-          undefined,
-          buildFilenameTemplateSourceOptions(sub)
-        );
-        const videoData = downloadResult?.videoData || downloadResult || {};
-        downloadedTwitchTitle = videoData.title || video.title;
-        twitchVideoDownloaded = true;
-
-        storageService.addDownloadHistoryItem({
-          id: uuidv4(),
-          title: downloadedTwitchTitle,
-          author: videoData.author || video.authorName || sub.author,
-          sourceUrl: video.url,
-          finishedAt: Date.now(),
-          status: "success",
-          videoPath: videoData.videoPath,
-          thumbnailPath: videoData.thumbnailPath,
-          videoId: videoData.id,
-          subscriptionId: sub.id,
-          platform: platformFromUrl(video.url),
-          sourceKind: "subscription",
-          totalSize:
-            typeof videoData.fileSize === "string" ||
-            typeof videoData.fileSize === "number"
-              ? String(videoData.fileSize)
-              : undefined,
-        });
-        newVideoCount += 1;
-
-        currentLastTwitchVideoId = video.id;
-        currentLastVideoLink = video.url;
-        currentDownloadCount += 1;
-
-        const updateResult = await db
-          .update(subscriptions)
-          .set({
-            lastTwitchVideoId: currentLastTwitchVideoId,
-            lastVideoLink: currentLastVideoLink,
-            downloadCount: currentDownloadCount,
-          })
-          .where(eq(subscriptions.id, sub.id))
-          .returning({ id: subscriptions.id });
-
-        if (updateResult.length === 0) {
-          logger.warn(
-            "Twitch subscription was deleted after download completed",
-            getSubscriptionLogContext(sub, { latestVideoUrl: video.url })
-          );
-          break;
-        }
-
-        notifySubscriptionDownloadResult({
-          taskTitle: downloadedTwitchTitle,
-          status: "success",
-          sourceUrl: video.url,
-        });
-      } catch (downloadError: any) {
-        const errorMessage = getErrorMessage(
-          downloadError,
-          "Download failed"
-        );
-
-        if (twitchVideoDownloaded) {
-          logger.error(
-            "Error updating Twitch subscription after video download",
-            downloadError,
-            getSubscriptionLogContext(sub, { latestVideoUrl: video.url })
-          );
-
-          notifySubscriptionDownloadResult({
-            taskTitle: downloadedTwitchTitle,
-            status: "fail",
-            sourceUrl: video.url,
-            error: `Subscription processing failed after download: ${errorMessage}`,
-          });
-          break;
-        }
-
-        logger.error(
-          "Error downloading Twitch subscription video",
-          downloadError,
-          getSubscriptionLogContext(sub, { latestVideoUrl: video.url })
-        );
-        storageService.addDownloadHistoryItem({
-          id: uuidv4(),
-          title: video.title || `Video from ${sub.author}`,
-          author: video.authorName || sub.author,
-          sourceUrl: video.url,
-          finishedAt: Date.now(),
-          status: "failed",
-          error: errorMessage,
-          subscriptionId: sub.id,
-          platform: platformFromUrl(video.url),
-          sourceKind: "subscription",
-        });
-        notifySubscriptionDownloadResult({
-          taskTitle: video.title || `Video from ${sub.author}`,
-          status: "fail",
-          sourceUrl: video.url,
-          error: errorMessage,
-        });
-        break;
-      }
-    }
-
-    return newVideoCount;
+    return processTwitchSubscriptionVideosImpl(sub, videosToProcess);
   }
 
   startScheduler() {
